Move header-related code to a separate file
